+++ driver/src/pm/xaie_reset_aie.c
@@ -84,7 +84,7 @@ static void _XAie_RstSetShimReset(XAie_DevInst *DevInst, XAie_LocType Loc,
 ******************************************************************************/
 AieRC _XAie_RstShims(XAie_DevInst *DevInst, u32 StartCol, u32 NumCols)
 {
-	for (u32 C = StartCol; C < (StartCol + NumCols); C++) {
+	for (u8 C = StartCol; C < (StartCol + NumCols); C++) {
 		XAie_LocType Loc = XAie_TileLoc(C, 0);
 
 		_XAie_RstSetShimReset(DevInst, Loc, XAIE_ENABLE);
@@ -96,7 +96,7 @@ AieRC _XAie_RstShims(XAie_DevInst *DevInst, u32 StartCol, u32 NumCols)
 	XAie_RunOp(DevInst, XAIE_BACKEND_OP_ASSERT_SHIMRST,
 			(void *)(uintptr_t)XAIE_DISABLE);
 
-	for (u32 C = StartCol; C < (StartCol + NumCols); C++) {
+	for (u8 C = StartCol; C < (StartCol + NumCols); C++) {
 		XAie_LocType Loc = XAie_TileLoc(C, 0);
 
 		_XAie_RstSetShimReset(DevInst, Loc, XAIE_DISABLE);
